move change collector into function, document interface
