+++ x-pack/plugin/transform/src/main/java/org/elasticsearch/xpack/transform/transforms/Function.java
@@ -10,7 +10,7 @@ import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.index.IndexRequest;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.client.Client;
-import org.elasticsearch.search.aggregations.AggregationBuilder;
+import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
 import org.elasticsearch.xpack.core.transform.transforms.SourceConfig;
 import org.elasticsearch.xpack.core.transform.transforms.TransformIndexerStats;
@@ -19,7 +19,83 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
+/**
+ * Interface for transform functions (e.g. pivot)
+ *
+ * The function interface abstracts:
+ *  - mapping deduction
+ *  - data preview
+ *  - validation
+ *  - collection of changes (finding the minimal update)
+ *  - querying the source index
+ *  - processing search results in order to write them to dest
+ *  - access to the cursor (for resilience and pausing a transform)
+ */
 public interface Function {
+
+    /**
+     * Change collector
+     *
+     * The purpose of the change collector is minimizing the update required for continuous transforms.
+     *
+     * In a nutshell it uses queries to
+     * 1. check and collect what needs to be updated
+     * 2. apply the collected changes as filter query.
+     */
+    public interface ChangeCollector {
+
+        boolean collectChanges(SearchResponse searchResponse);
+
+        QueryBuilder filterByChanges(long lastCheckpointTimestamp, long nextcheckpointTimestamp);
+
+        SearchSourceBuilder buildChangesQuery(SearchSourceBuilder sourceBuilder, Map<String, Object> position, int pageSize);
+
+        void clear();
+
+        Map<String, Object> getBucketPosition();
+    }
+
+    /**
+     * Deduce mappings based on the input mappings and the known configuration.
+     *
+     * @param client a client instance for querying the source mappings
+     * @param sourceConfig the source configuration
+     * @param listener listener to take the deduced mapping
+     */
+    void deduceMappings(Client client, SourceConfig sourceConfig, ActionListener<Map<String, String>> listener);
+
+    void preview(
+        Client client,
+        Map<String, String> headers,
+        SourceConfig sourceConfig,
+        Map<String, String> fieldTypeMap,
+        int numberOfBuckets,
+        ActionListener<List<Map<String, Object>>> listener
+    );
+
+    /**
+     * Validate configuration.
+     *
+     * @param listener the result listener
+     */
+    void validateConfig(ActionListener<Boolean> listener);
+
+    /**
+     * Runtime validation by querying the source and checking if source and config fit.
+     *
+     * @param client a client instance for querying the source
+     * @param sourceConfig the source configuration
+     * @param listener the result listener
+     */
+    void validateQuery(Client client, SourceConfig sourceConfig, ActionListener<Boolean> listener);
+
+    /**
+     * Create a change collector instance and return it
+     *
+     * @param synchronizationField the field used for synchronizing (continuous mode)
+     * @return a change collector instance
+     */
+    ChangeCollector buildChangeCollector(String synchronizationField);
 
     /**
      * Get the initial page size for this function.
@@ -34,11 +110,15 @@ public interface Function {
      */
     int getInitialPageSize();
 
+    /**
+     * Build the query for the next iteration
+     *
+     * @param builder a searchsource builder instance
+     * @param position current position (cursor/page)
+     * @param pageSize the pageSize, defining how much data to request
+     * @return the searchSource, expanded with the relevant parts
+     */
     SearchSourceBuilder source(SearchSourceBuilder builder, Map<String, Object> position, int pageSize);
-
-    AggregationBuilder aggregation(Map<String, Object> position, int pageSize);
-
-    ChangeCollector buildChangeCollector(String synchronizationField);
 
     boolean supportsIncrementalBucketUpdate();
 
@@ -51,19 +131,4 @@ public interface Function {
     );
 
     Map<String, Object> getAfterKey(SearchResponse searchResponse);
-
-    void validateQuery(Client client, SourceConfig sourceConfig, ActionListener<Boolean> listener);
-
-    void validateConfig(ActionListener<Boolean> listener);
-
-    void deduceMappings(Client client, SourceConfig sourceConfig, ActionListener<Map<String, String>> listener);
-
-    void preview(
-        Client client,
-        Map<String, String> headers,
-        SourceConfig sourceConfig,
-        Map<String, String> fieldTypeMap,
-        int numberOfBuckets,
-        ActionListener<List<Map<String, Object>>> listener
-    );
 }
